Changed int to long long

+++ src/Card.h
@@ -4,8 +4,8 @@
 #include <compare>
 
 struct Card {
-    int suit;
-    int number;
+    long long suit;
+    long long number;
 
     auto operator<=>(const Card& rhs) const { return number <=> rhs.number; }
 };
