+++ src/DeckManager.cpp
@@ -3,7 +3,7 @@
 #include <algorithm>
 #include <random>
 
-DeckManager::DeckManager(const int cards_of_same_suit_count)
+DeckManager::DeckManager(const long long cards_of_same_suit_count)
     : mersenne_twister_engine_(std::random_device{}()),
       current_card_index_(0),
       cards_of_same_suit_count_(cards_of_same_suit_count) {
@@ -17,8 +17,8 @@ DeckManager::DeckManager(const int cards_of_same_suit_count)
 
 void DeckManager::GenerateDeck() {
     for (int i = 0; i < deck_size_; i++) {
-        const int card_suit = i / cards_of_same_suit_count_;
-        const int card_number = i % cards_of_same_suit_count_;
+        const long long card_suit = i / cards_of_same_suit_count_;
+        const long long card_number = i % cards_of_same_suit_count_;
         const Card card(card_suit, card_number);
         deck_[i] = card;
     }
